refactor(codex): the CLI surface moves into the package
Phase 2b, step three — the first agent to actually leave the app. `apps/cli/src/agents/codex/`
was 7 files and ~606 lines; it is one 18-line shim now.

The split follows the plan: the package exports a `codexCliSpec` DESCRIPTOR and
the CLI builds the commander tree from it with `buildAgentCommand`, which stays
in the app because its closure is the whole create/attach pipeline. Moving the
factory would have dragged ~1,500 lines out for no gain.

`@agentbox/agent-codex/cli` is a second entry point rather than the package
root. The root is the docker/cloud behaviour that `sandbox-docker` and
`sandbox-cloud` receive, and it must stay importable by things with no business
pulling in commander and the prompt stack.

Both literal-import tables keep literal specifiers
(`import('@agentbox/agent-codex/cli')`) — a computed one resolves in the dev
tree and MODULE_NOT_FOUNDs only in the published bundle, so this was verified
with a packed install into a clean prefix, not just `pnpm build`.

One test had to move with the code, and the reason is the recurring one: the
CLI's `codexAuthAvailable` test mocked `@agentbox/agent-codex`, which worked
only while the IMPORTER was outside the package. `cli/host-creds.ts` now reaches
its sibling `../docker-sync.js` directly, so only a test inside the package can
intercept it. It lives in `packages/agent-codex/test/` and still fails when the
volume-probe fallback is removed.
