+++ apps/cli/test/login-binding.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { claudeLoginBinding } from '../src/agents/claude/login-binding.js';
-import { codexLoginBinding } from '../src/agents/codex/login-binding.js';
+import { codexLoginBinding } from '@agentbox/agent-codex/cli';
 import { opencodeLoginBinding } from '../src/agents/opencode/login-binding.js';
 import { withLoginDefaults } from '@agentbox/cli-kit';
