+++ apps/cli/test/skip-permissions.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { applySkipPermissions, type SkipPermissionsRule } from '@agentbox/cli-kit';
 import { claudeRuntime } from '../src/agents/claude/runtime.js';
-import { codexRuntime } from '../src/agents/codex/runtime.js';
+import { codexRuntime } from '@agentbox/agent-codex/cli';
 import { opencodeRuntime } from '../src/agents/opencode/runtime.js';
 import type { EffectiveConfig } from '@agentbox/config';
