+++ packages/agent-codex/src/cli/module.ts
@@ -1,4 +1,4 @@
-import { defineAgentModule, type AgentModule } from '../index.js';
+import { defineAgentModule, type AgentModule } from '@agentbox/cli-kit';
 import { CODEX_LOGIN_SPEC } from './login.js';
 import { codexRuntime } from './runtime.js';
 import { resolveCodexTeleport } from './teleport.js';
