+++ packages/agent-codex/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/cli/index.ts'],
   format: ['esm'],
   target: 'node20',
   clean: true,
